requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.5.0
djangorestframework==3.14.0
gunicorn==23.0.0
mysqlclient==2.2.5
packaging==24.1
pillow==11.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.2
